S10P22A407-6 feat: Add naver login for login

+++ backend/api/src/main/java/store/dalkak/api/global/oauth/dto/GoogleUserAuthDto.java
@@ -14,7 +14,7 @@ import lombok.*;
 public class GoogleUserAuthDto {
     //access token, refresh token, id token
     String accessToken;
-    int expiresIn;
+    Long expiresIn;
     @NotNull
     String idToken;
     String scope;
