+++ backend/api/src/main/java/store/dalkak/api/global/oauth/dto/KakaoUserAuthDto.java
@@ -17,8 +17,8 @@ public class KakaoUserAuthDto {
     String refreshToken;
     @NotNull
     String idToken;
-    int expiresIn;
+    Long expiresIn;
     String scope;
-    int refreshTokenExpiresIn;
+    Long refreshTokenExpiresIn;
 
 }
